rename package from url -> pkgurl
