+++ app/scaffold/pkgs/pkgurl/url_test.go
@@ -1,4 +1,4 @@
-package url
+package pkgurl
 
 import (
 	"testing"
